# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for bitcoin operations.",
        "title": "BTC-CRYPTO-MANAGER",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/balance/{address}": {
            "get": {
                "description": "Retrieves balance information for a given Bitcoin address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bitcoin"
                ],
                "summary": "Find balance per Bitcoin address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bitcoin Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bitcoin Balance",
                        "schema": {
                            "$ref": "#/definitions/domain.BalanceDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/details/{address}": {
            "get": {
                "description": "Retrieves detailed information for a given Bitcoin address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bitcoin"
                ],
                "summary": "Find details per Bitcoin address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bitcoin Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bitcoin Details",
                        "schema": {
                            "$ref": "#/definitions/domain.DetailsAddress"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/send": {
            "post": {
                "description": "Creates a UTXO set for a given Bitcoin address and amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bitcoin"
                ],
                "summary": "Mount UTXO",
                "parameters": [
                    {
                        "description": "Bitcoin Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BitcoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UTXO",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.UTXOs"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/tx/{tx}": {
            "get": {
                "description": "Retrieves detailed information for a given Bitcoin transaction ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bitcoin"
                ],
                "summary": "Find details per transaction ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "tx",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction Details",
                        "schema": {
                            "$ref": "#/definitions/domain.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.HTTPErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Address": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "domain.BalanceDetail": {
            "type": "object",
            "properties": {
                "confirmed": {
                    "type": "string"
                },
                "unconfirmed": {
                    "type": "string"
                }
            }
        },
        "domain.BitcoinRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "string"
                }
            }
        },
        "domain.BitcoinReturnSend": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "txid": {
                    "type": "string"
                }
            }
        },
        "domain.DetailsAddress": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                },
                "balance_details": {
                    "$ref": "#/definitions/domain.BalanceDetail"
                },
                "total": {
                    "$ref": "#/definitions/domain.Total"
                },
                "totalTx": {
                    "type": "integer"
                }
            }
        },
        "domain.HTTPErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "errorMessage": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Total": {
            "type": "object",
            "properties": {
                "received": {
                    "type": "string"
                },
                "sent": {
                    "type": "string"
                }
            }
        },
        "domain.Transaction": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Address"
                    }
                },
                "block": {
                    "type": "integer"
                },
                "txID": {
                    "type": "string"
                }
            }
        },
        "domain.UTXOs": {
            "type": "object",
            "properties": {
                "utxos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BitcoinReturnSend"
                    }
                }
            }
        }
    }
}